Fix revtr nil pointer
When an WS request came in for an existing reverse traceroute output
channel, the oc pointer is never set to the channel.

Also, the map of of output channels to ids is never cleared when a
reverse traceroute completes. Added a complete callback to the
outputChan struct that is called when the output channel closes.